Ops Events - Debug
- Ops Events deploy, Fix list limit
- Parse only valid contracts
- Fix helpers.OpenseaNftEvent large quantities wrapper

+++ modules/helpers/opensea.go
@@ -38,7 +38,7 @@ type OpenseaNftEvent struct {
 	Chain           *string                 `json:"chain"`
 	ClosingDate     *int                    `json:"closing_date"`
 	Nft             *OpenseaNftAsset        `json:"nft"`
-	Quantity        *int                    `json:"quantity"`
+	Quantity        *float64                `json:"quantity"`
 	Seller          *string                 `json:"seller"`
 	Buyer           *string                 `json:"buyer"`
 	Payment         *OpenseaNftEventPayment `json:"payment"`
